feat: delete student data

+++ controller/student.js
@@ -95,6 +95,7 @@ exports.deleteStudent = async(req, res) => {
   try{
     if(!req.user.isAdmin) return responseHandler(res, 403, 'Unauthorized');
     const { id } = req.params;
+    if(!id) return responseHandler(res, 400, 'Required: id');
     const studentData = await Student.findByPk(parseInt(id));
     if(!studentData) return responseHandler(res, 404, 'Data not found');
     await Student.destroy({
